moving unit tests for calibrate to their own file
